Enhance (sign in with google) code.
